Use lxc-attach to run commands inside container
Instead of using ssh/scp, we can use lxc-attach to attach to a running
container and execute a single command.

This means that we don't have to wait for the container to get an IP
address. Finding the IP address in a robust way seems non-trivial: the
script was looking at /var/lib/misc/dnsmasq.leases to determine it,
but that file does not exist when I run LXC on Debian.