Consolidate jenkins scripts
The jenkins scripts are all pretty generic, with the exception of the
zerovm-cli script, which must copy test reports out of the lxc before
continuing to the packaging phase.

This does not include toolchain; we don't even have a stable
compile/build status at the moment for toolchain, so this still needs a
lot of work.