toolchain: overwrite ~/.bashrc instead of prepending
We would like to set ZRT_ROOT when the generic package.sh script is
run. The build commands are executed SSH, which means that we have a
non-login non-interactive shell. Bash (the shell on the build machine)
would then normally not execute ~/.bashrc or ~/.bash_profile. However,
Bash has code to detect when it's started by sshd and will then
execute ~/.bashrc despite the session being non-interactive.

Ubuntu ships with an elaborate default .bashrc file, which among other
things exit early if it sees that $PS1 is empty. An empty $PS1 is a
sign that the session is non-interactive. This defeats the purpose of
Bash executing the ~/.bashrc file!

Since we don't want or need the Ubuntu defaults, so we just overwrite
the ~/.bashrc file. This simplifies the whole thing and removes
assumptions about the Ubuntu default ~/.bashrc file.