packager.py: fix tar archive file collection
Previously, we were creating tarballs for packaging by including * and
excluding debian/. This works for many cases, but causes errors if a
project repo contains files beginning with ".". `debuild` will fail in
this case, resulting in a "local changes detected" error (because there
is a `.whatever` file in the local clone which is not present in the
tarball).

This change fixes that by add more explicit file inclusion glob patterns
and --exclude directives.