buildenv.sh: fix the way we get the ip from lxcs
Instead of scraping the IP from the dnsmasq.leases file, just try to get
the IP by using `lxc-ls --fancy`.

This seems to be a cleaner and more stable way to get the IP. Also, it
seems that the old method stopped working and was causing build jobs to
fail, due to a failure to get an IP. Through some experiementation on
the build server, I found that the dnsmasq.leases file is no longer
populated.

I'm not sure what happened, to be honest. The only major thing that
changed is that we installed supybot on the same box (for IRC logging,
etc.), but I don't see any other updates that could have caused this
issue.